requirements.txt
amqp==5.0.5
asgiref==3.2.10
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.1.1
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.30.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
jdcal==1.4.1
kombu==5.0.2
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.6
packaging==20.9
pandas==1.2.2
Pillow==8.0.1
prompt-toolkit==3.0.16
protobuf==3.15.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==1.3.7
xlwt==1.3.0
